✨ (cmd): update range handling to support float64 for chapter numbers 📝 (grabber): improve comments for clarity and consistency 🔧 (http): enhance request function with TLS configuration 🔄 (ranges): refactor parsing logic for chapter ranges to support decimals
Refactor the code to improve clarity and functionality. The changes enhance the handling of chapter ranges, allowing for decimal values, and improve the overall documentation for better understanding.

+++ src/cmd/root.go
@@ -107,7 +107,7 @@ func Run(cmd *cobra.Command, args []string) {
 			fmt.Println(color.YellowString("Canceled by user"))
 			os.Exit(0)
 		}
-		rngs = []ranges.Range{{Begin: 1, End: int64(lastChapter)}}
+		rngs = []ranges.Range{{Begin: 1.0, End: lastChapter}}
 	} else {
 		settings.Range = getRangesArg(args)
 		rngs, err = ranges.Parse(settings.Range)
